% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{beta_dist}
\alias{beta_dist}
\title{Sample Beta Distribution and Return Density}
\usage{
beta_dist(mean, variance, samples = 1000, variance_factor = 0.1)
}
\arguments{
\item{mean}{Numeric scalar. Desired mean of the beta distribution. Must be between 0 and 1 (exclusive).}

\item{variance}{Numeric scalar. Desired variance of the beta distribution. Will be constrained by \code{variance_factor}.}

\item{samples}{Integer. Number of samples to draw from the beta distribution. Default is 1000.}

\item{variance_factor}{Numeric scalar. Limits the allowed variance as a fraction of the theoretical maximum variance. Default is 0.1.}
}
\value{
A tibble with columns \code{x} and \code{y} representing the smoothed density estimate of the sampled beta distribution.
}
\description{
Generates a beta distribution based on a specified mean and variance,
and returns a density estimate as a data frame.
}
\details{
The shape (i.e., skew and peak) of a Beta distribution is driven by the ratio alpha / (alpha + beta) (which controls the mean), and the sum alpha + beta (which controls the concentration/peakedness).
}
\examples{
beta_dist(mean = 0.4, variance = 0.01)

}
