% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reefspawn-data.R
\docType{data}
\name{polyp_density_data}
\alias{polyp_density_data}
\title{Load and augment coral polyp density data}
\format{
An object of class \code{data.frame} with 45 rows and 7 columns.
}
\usage{
polyp_density_data
}
\value{
A \code{polyp_density_data} data frame with all species and assigned growthforms.
}
\description{
Loads \code{polyp_density.csv}, adds an estimated entry for \emph{Acropora millepora}
based on similar species (\emph{A. nasuta} and \emph{A. spathulata}), and assigns growthform classifications.
}
\details{
\itemize{
\item \verb{Acropora millepora} is not present in the original dataset.
\item To estimate it, data for \emph{A. nasuta} and \emph{A. spathulata} are duplicated and relabeled.
\item A new \code{spp} code "AM" is assigned to the added entries.
\item Growthforms are recoded using \code{fct_recode()} based on species name.
}
}
\examples{
head(polyp_density_data)

}
\keyword{datasets}
