% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_coral_predict.R
\name{posterior_coral_predict}
\alias{posterior_coral_predict}
\title{Generate Posterior Predictions of Coral Colony Area}
\usage{
posterior_coral_predict(
  brms,
  ndraws = 10000,
  seed = NULL,
  newdata,
  correct_species = FALSE
)
}
\arguments{
\item{brms}{A fitted brms model predicting colony width.}

\item{ndraws}{Integer. Number of posterior draws to sample.}

\item{newdata}{A data frame containing the predictor variables for prediction.}

\item{correct_species}{Logical. Whether to rename "Acropora cf. digitifera" to "Acropora digitifera" in predictions.}
}
\value{
A tibble with posterior draws in long format, including species, width, area, and cumulative area.
}
\description{
This function uses a fitted \code{brms} model to generate posterior predictions of colony size (width),
transforms width to area, and returns cumulative area predictions per species in long format.
}
