% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reefspawnfunctions.R
\name{predict_colony_width}
\alias{predict_colony_width}
\title{Predict colony width using posterior predict (lognormal model)}
\usage{
predict_colony_width(newdata, draws, draw_id = 1)
}
\arguments{
\item{newdata}{A data.frame with a \code{species} column.}

\item{draws}{Posterior draws from \code{posterior_coeffs_sizedistribution}.}

\item{draw_id}{Integer index of the draw to use.}
}
\value{
Numeric vector of predicted widths.
}
\description{
Predict colony width using posterior predict (lognormal model)
}
\examples{
\dontrun{
newdata <- data.frame(species = c("Acropora", "Porites"))
predict_colony_width(newdata, posterior_coeffs_sizedistribution, draw_id = 10)
}
}
