% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reefspawnfunctions.R
\name{predict_oocytes}
\alias{predict_oocytes}
\title{Predict oocytes per reproductive polyp using posterior predict (ZIP model)}
\usage{
predict_oocytes(newdata, draws, draw_id = 1)
}
\arguments{
\item{newdata}{A data.frame with \code{species} and \code{width} columns.}

\item{draws}{Posterior draws from \code{posterior_coeffs_oocyte_output}.}

\item{draw_id}{Integer index of the draw to use.}
}
\value{
Numeric vector of expected counts.
}
\description{
Predict oocytes per reproductive polyp using posterior predict (ZIP model)
}
\examples{
\dontrun{
newdata <- expand.grid(species = levels(coralsize$species), width = seq(0,50,5))
newdata$oocytes <- predict_oocytes(newdata, posterior_coeffs_oocyte_output, draw_id = 100)
}
}
