% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reefspawnfunctions.R
\name{predict_repro_prob}
\alias{predict_repro_prob}
\title{Predict reproductive probability using posterior predict (bernoulli model)}
\usage{
predict_repro_prob(newdata, draws, draw_id = 1)
}
\arguments{
\item{newdata}{A data.frame with \code{species} and \code{width} columns.}

\item{draws}{Posterior draws from \code{posterior_coeffs_reproductive_probability}.}

\item{draw_id}{Integer index of the draw to use.}
}
\value{
Numeric vector of probabilities.
}
\description{
Predict reproductive probability using posterior predict (bernoulli model)
}
\examples{
\dontrun{
newdata <- expand.grid(species = levels(coralsize$species), width = seq(0,50,5))
newdata$reproductive_probability <- predict_repro_prob(newdata, posterior_coeffs_reproductive_probability, draw_id = 20)
}
}
