% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reefspawn-package.R
\docType{package}
\name{reefspawn-package}
\alias{reefspawn}
\alias{reefspawn-package}
\title{reefspawn: Community scale simulations of coral reproductive output}
\description{
Functions to scale coral reproductive output and spawning in Indo-Pacific corals
}
\seealso{
Useful links:
\itemize{
  \item \url{https://marine-ecologist.github.io/reefspawn}
}

}
\author{
\strong{Maintainer}: George Roff \email{george.roff@csiro.au}

}
\keyword{internal}
