% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setplot.R
\name{setplot}
\alias{setplot}
\title{Generate a Rectangular sf Plot Grid}
\usage{
setplot(length, width, crs = 3857, quiet = TRUE)
}
\arguments{
\item{length}{Numeric. Length of the plot in meters (X-dimension).}

\item{width}{Numeric. Width of the plot in meters (Y-dimension).}

\item{crs}{Integer. Coordinate Reference System (default = 3857).}
}
\value{
An \code{sf} object representing a grid of 1m x 1m cells covering the defined plot.
}
\description{
Creates a simple rectangular spatial plot of specified \code{length} and \code{width} in meters,
and returns a 1m x 1m \code{sf} grid overlaid on that area. CRS is EPSG:3857 by default.
}
\examples{
grid <- setplot(10, 5)
plot(grid)

}
